requirements.txt
streamlit
upstage
python-dotenv
pandas
langchain-upstage
langchain-core 
langchain-community 
langchain-text-splitters 
faiss-cpu
easyocr